server/requirements.txt
anyio==3.6.2
attrs==22.1.0
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
dateparser==1.1.4
exceptiongroup==1.0.4
fastapi==0.86.0
h11==0.14.0
idna==3.4
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.2
pdfminer.six==20221105
platformdirs==2.5.4
pluggy==1.0.0
pycparser==2.21
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.12
uvicorn==0.19.0
